refactor: item component
