+++ src/routes/Settings/URLsManager/Item/index.ts
@@ -1,5 +1,9 @@
 // Copyright (C) 2017-2024 Smart code 203358507
 
 import Item from './Item';
+import AddItem from './AddItem';
 
-export default Item;
+export {
+    Item,
+    AddItem,
+}
