+++ src/routes/Settings/URLsManager/URLsManager.less
@@ -25,6 +25,17 @@
         gap: 1rem;
         padding: 1.5rem 0;
     }
+    
+    .item {
+        display: flex;
+        padding: 1rem 1.5rem;
+        border-radius: var(--border-radius);
+        transition: 0.3s all ease-in-out;
+        background-color: transparent;
+        border: 2px solid transparent;
+        justify-content: space-between;
+        position: relative;
+    }
 
     .footer {
         display: flex;
